read_habitatmap_stdized(): temporary hack to fix CRS *
It appears that the raw data source and hence, the data source habitatmap_stdized, have a CRS with approximately (not entirely exactly) the proj4string of the Belge Lambert 72 (EPSG code 31300), and *not* Belgian Lambert 72 (EPSG code 31370) which should be taken. Consequently it also lacks an EPSG code to define the CRS.

This gives problems in further spatial processing.

While this will need to be solved in a next version of the processed data source, currently the reading function does the CRS transformation so this is a temporary hack.